--- modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-functions/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/functions/IsEmailAddressFunction.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.dynamic.data.mapping.form.evaluator.functions;
-
-import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunction;
-import com.liferay.portal.kernel.util.Validator;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Leonardo Barros
- */
-@Component(
-	immediate = true,
-	property = "ddm.form.evaluator.function.name=isEmailAddress",
-	service = DDMExpressionFunction.class
-)
-public class IsEmailAddressFunction implements DDMExpressionFunction {
-
-	@Override
-	public Object evaluate(Object... parameters) {
-		if (parameters.length != 1) {
-			throw new IllegalArgumentException("One parameter is expected");
-		}
-
-		return Validator.isEmailAddress(parameters[0].toString());
-	}
-
-}
