+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/functions/IsEmptyFunction.java
@@ -1,0 +1,60 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions;
+
+import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunction;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.util.stream.Stream;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Leonardo Barros
+ */
+@Component(
+	immediate = true, property = "ddm.form.evaluator.function.name=isEmpty",
+	service = DDMExpressionFunction.class
+)
+public class IsEmptyFunction implements DDMExpressionFunction {
+
+	@Override
+	public Object evaluate(Object... parameters) {
+		if (parameters == null) {
+			return true;
+		}
+
+		if ((parameters.length == 1) && isArray(parameters[0])) {
+			Object[] values = (Object[])parameters[0];
+
+			if (values.length == 0) {
+				return true;
+			}
+
+			Stream<Object> stream = Stream.of(values);
+
+			return !stream.anyMatch(Validator::isNotNull);
+		}
+
+		return Validator.isNull(parameters[0]);
+	}
+
+	protected boolean isArray(Object parameter) {
+		Class<?> clazz = parameter.getClass();
+
+		return clazz.isArray();
+	}
+
+}
